Refactor per correggere alcuni errori di codice. corretto funzionamento creazione rete. spostate soglie peso nel package RetrieveData

+++ src/main/java/server/database2/MessageMedicPatientDB.java
@@ -107,8 +107,8 @@ public class MessageMedicPatientDB {
         try {
             conn = DBConnectOnline.getInstance().getConnection();
             PreparedStatement st = conn.prepareStatement(sql);
-            st.setInt(1, ((Double) map.get("medic_id")).intValue());
-            st.setInt(2, ((Double) map.get("patient_id")).intValue());
+            st.setInt(1, (Integer) map.get("medic_id"));
+            st.setInt(2, (Integer) map.get("patient_id"));
             st.setString(3, String.valueOf(map.get("subject")));
             st.setString(4, String.valueOf(map.get("timedate")));
             st.setBoolean(5, (Boolean) map.get("medic_sender"));
